Products controller added
